misc/SerialCord/requirements.txt
discord.py==1.3.2
pexpect==4.8.0
